fixup! Append (1) to the first of multiple tests
Before: on multiple invocations of the same test, only the
second and subsequent would get (2), (3), etc.

Now: even the first one now gets (1). This is slightly cleaner,
in that the viewer now has an easier time locating the first
invocation of a multi-run test; it's also clearer when skimming
in order that a given test will be run again.