fixup! Workaround for horrible jenkins /run bug
Jenkins is intercepting URLs ending in /run, redirecting
them to the test index page, making it impossible to view
results of the docker_cli.run test.

I've spent over an hour looking at jenkins plugins and config,
found nothing useful. At this point, disgusting as it may be,
the best ROI is to kludge a workaround.