Experimental: ignore failures in known-problem subtests
  https://trello.com/c/qqlVHKRX/353-5-silence-known-ci-failures

A common pattern seems to be:

  1) New docker introduces a regression
  2) Wait months for docker to fix
  3) Wait more months for new RPM build

It is frustrating to see those failures over and over on each
test run.

This commit introduces a mechanism for associating a docker build
(exact NVR), a subtest, and a comment describing the known failure.
If we see this failure, we treat it as a skipped test. Bonus: if
we see a failure in a subtest, and there's a known-failure for
a different NVR but same NV (e.g. docker-1.12.5-6 but not -7),
emit a warning near the failure message; this can help a
maintainer update the known-failures list.

Problems this solves:

  * constant failures in ADEPT nightly runs with same NVR
  * constant failures in ADEPT runs on github PRs

Problems this does not solve:

  * ADEPT runs on new builds which do not address known failures
  * ADEPT runs on docker PRs which "  "    " "    " "    " "
    (theoretical concern only: that is still just a pipe dream)

Signed-off-by: Ed Santiago <santiago@redhat.com>