Add class to represent canonical distro. names
It's desireable to have a single, simple set of
distribution names to act upon.  However,
occasionally the detection methods can change.
This class behaves like a regular string for normal use,
but the allows methods to be added over time to support
new and/or overlaping types of detection.

In additon, because the parsing results are preserved,
it supports future additions of supporting methods
to provide more detailed information.  For example
if a distro. was detected as "RHEL", it could also support
a ``full_name()`` call to return "RHEL Atomic Host".

Signed-off-by: Chris Evich <cevich@redhat.com>