Template for creating new tests, cheaply
This provides a new bash-script scaffolding that developers
can grab, modify, and submit to QE for inclusion in autotest.
The scaffolding provides logging, failure checks, helper
functions for testing command output, and various examples.

The idea is to publish this in an easily-accessible location,
possibly part of the Trello QE checklist. Developers can
grab this as part of a bug/feature fix, quickly add a
sanity test for their code, and submit the script to QE.
QE can then use as-is or, if warranted, rewrite into
the docker-autotest framework.

Initial commit is entirely docker-centric because docker is
the only tool we currently test. We need to have some
conversations about that.

Signed-off-by: Ed Santiago <santiago@redhat.com>